Refactoring
Renamed namespace ScientificReport.DAL.Entities -> ScientificReport.DAL.Entities.UserProfile in ScientificReport.DAL/Entities/UserProfile
