+++ ScientificReport.BLL/Interfaces/IPublicationService.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using ScientificReport.DAL.Entities;
+using ScientificReport.DAL.Entities.UserProfile;
 
 namespace ScientificReport.BLL.Interfaces
 {
