+++ ScientificReport.BLL/Interfaces/IUserProfileService.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Threading.Tasks;
 using Microsoft.AspNetCore.Identity;
 using ScientificReport.DAL.Entities;
-using ScientificReport.DAL.Roles;
+using ScientificReport.DAL.Entities.UserProfile;
 
 namespace ScientificReport.BLL.Interfaces
 {
