+++ ScientificReport.BLL/Services/PublicationService.cs
@@ -4,6 +4,7 @@ using System.Linq;
 using ScientificReport.BLL.Interfaces;
 using ScientificReport.DAL.DbContext;
 using ScientificReport.DAL.Entities;
+using ScientificReport.DAL.Entities.UserProfile;
 using ScientificReport.DAL.Repositories;
 
 namespace ScientificReport.BLL.Services
