+++ ScientificReport.BLL/Services/UserProfileService.cs
@@ -6,6 +6,7 @@ using Microsoft.AspNetCore.Identity;
 using ScientificReport.BLL.Interfaces;
 using ScientificReport.DAL.DbContext;
 using ScientificReport.DAL.Entities;
+using ScientificReport.DAL.Entities.UserProfile;
 using ScientificReport.DAL.Repositories;
 
 namespace ScientificReport.BLL.Services
