+++ ScientificReport.DAL/Entities/ApplicantsPatentLicenseActivities.cs
@@ -9,7 +9,7 @@ namespace ScientificReport.DAL.Entities
 		public Guid Id { get; set; }
 		
 		public string ApplicantId { get; set; }
-		public virtual UserProfile Applicant { get; set; }
+		public virtual UserProfile.UserProfile Applicant { get; set; }
 		
 		public int PatentLicenseActivityId { get; set; }
 		public virtual PatentLicenseActivity PatentLicenseActivity { get; set; }
