+++ ScientificReport.DAL/Entities/AuthorsPatentLicenseActivities.cs
@@ -9,7 +9,7 @@ namespace ScientificReport.DAL.Entities
 		public Guid Id { get; set; }
 		
 		public int AuthorId { get; set; }
-		public virtual UserProfile Author { get; set; }
+		public virtual UserProfile.UserProfile Author { get; set; }
 		
 		public int PatentLicenseActivityId { get; set; }
 		public virtual PatentLicenseActivity PatentLicenseActivity { get; set; }
