+++ ScientificReport.DAL/Repositories/UserProfilesScientificWorksRepository.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using Microsoft.EntityFrameworkCore;
 using ScientificReport.DAL.DbContext;
-using ScientificReport.DAL.Entities;
+using ScientificReport.DAL.Entities.UserProfile;
 using ScientificReport.DAL.Interfaces;
 
 namespace ScientificReport.DAL.Repositories
