+++ ScientificReport.DTO/Models/Department/DepartmentEditModel.cs
@@ -6,9 +6,9 @@ namespace ScientificReport.DTO.Models.Department
 	{
 		public DAL.Entities.Department Department { get; set; }
 		
-		public DAL.Entities.UserProfile Head { get; set; }
+		public DAL.Entities.UserProfile.UserProfile Head { get; set; }
 		
-		public IEnumerable<DAL.Entities.UserProfile> Users { get; set; }
+		public IEnumerable<DAL.Entities.UserProfile.UserProfile> Users { get; set; }
 		
 		public IEnumerable<DAL.Entities.ScientificWork> ScientificWorks { get; set; }
