+++ ScientificReport.DTO/Models/ScientificWorks/ScientificWorksDetails.cs
@@ -6,6 +6,6 @@ namespace ScientificReport.DTO.Models.ScientificWorks
 	public class ScientificWorksDetails
 	{
 		public ScientificWork ScientificWork { get; set; }
-		public ICollection<DAL.Entities.UserProfile> Authors { get; set; }
+		public ICollection<DAL.Entities.UserProfile.UserProfile> Authors { get; set; }
 	}
 }
