+++ ScientificReport.DTO/Models/ScientificWorks/ScientificWorksEdit.cs
@@ -6,7 +6,7 @@ namespace ScientificReport.DTO.Models.ScientificWorks
 	public class ScientificWorksEdit
 	{
 		public ScientificWork ScientificWork { get; set; }
-		public IEnumerable<DAL.Entities.UserProfile> Authors { get; set; }
-		public IEnumerable<DAL.Entities.UserProfile> Users { get; set; }
+		public IEnumerable<DAL.Entities.UserProfile.UserProfile> Authors { get; set; }
+		public IEnumerable<DAL.Entities.UserProfile.UserProfile> Users { get; set; }
 	}
 }
