+++ ScientificReport.DTO/Models/UserProfile/UserProfileEditModel.cs
@@ -1,11 +1,10 @@
 using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
 
 namespace ScientificReport.DTO.Models.UserProfile
 {
 	public class UserProfileEditModel
 	{
-		public DAL.Entities.UserProfile UserProfile { get; set; }
+		public DAL.Entities.UserProfile.UserProfile UserProfile { get; set; }
 		
 		public IEnumerable<DAL.Roles.UserProfileRole> AllRoles { get; set; }
