+++ ScientificReport/Controllers/TeacherReportController.cs
@@ -1,13 +1,12 @@
 using System;
-using System.Collections;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using ScientificReport.BLL.Services;
 using ScientificReport.DAL.DbContext;
-using ScientificReport.DAL.Entities;
 using ScientificReport.DAL.Entities.Reports;
+using ScientificReport.DAL.Entities.UserProfile;
 
 namespace ScientificReport.Controllers
 {
