+++ ScientificReport/ScientificReport.csproj
@@ -52,10 +52,6 @@
 	</ItemGroup>
 
 	<ItemGroup>
-	  <Folder Include="Migrations" />
-	</ItemGroup>
-
-	<ItemGroup>
 	  <_ContentIncludedByDefault Remove="Views\ScientificWork\Create.cshtml" />
 	  <_ContentIncludedByDefault Remove="Views\ScientificWork\Delete.cshtml" />
 	  <_ContentIncludedByDefault Remove="Views\ScientificWork\Details.cshtml" />
