+++ ScientificReport/Views/Shared/_LoginPartial.cshtml
@@ -1,5 +1,5 @@
 @using Microsoft.AspNetCore.Identity
-@using ScientificReport.DAL.Entities
+@using ScientificReport.DAL.Entities.UserProfile
 @inject SignInManager<UserProfile> SignInManager
 @inject UserManager<UserProfile> UserManager
