+++ ScientificReport/Views/UserProfile/Details.cshtml
@@ -1,4 +1,4 @@
-@model ScientificReport.DAL.Entities.UserProfile
+@model ScientificReport.DAL.Entities.UserProfile.UserProfile
 
 @{
 	ViewData["Title"] = "Details";
