+++ ScientificReport/Views/UserProfile/Index.cshtml
@@ -1,4 +1,4 @@
-@model IEnumerable<ScientificReport.DAL.Entities.UserProfile>
+@model IEnumerable<ScientificReport.DAL.Entities.UserProfile.UserProfile>
 
 @{
 	ViewData["Title"] = "Index";
@@ -41,9 +41,9 @@
 				@Html.DisplayFor(modelItem => item.Email)
 			</th>
 			<td>
-				<a asp-action="Edit" asp-route-id="@item.Id">Edit</a> |
-				<a asp-action="Details" asp-route-id="@item.Id">Details</a> |
-				<form asp-action="Delete" asp-route-id="@item.Id" method="post">
+				<a asp-action="Details" asp-route-id="@item.Id" class="btn btn-sm btn-light">Details</a>
+				<a asp-action="Edit" asp-route-id="@item.Id" class="btn btn-sm btn-success">Edit</a>
+				<form asp-action="Delete" asp-route-id="@item.Id" method="post" style="display: inline;">
 					<button type="submit" class="btn btn-sm btn-danger">Delete</button>
 				</form>
 			</td>
